エラー及びバリデーションメッセージをpropertiesにまとめました。管理者権限での削除機能をついかしました。

+++ practice-2/src/main/java/com/example/demo/controller/admin/AdminListController.java
@@ -3,16 +3,20 @@ package com.example.demo.controller.admin;
 import jakarta.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import com.example.demo.entity.AdminEntity;
 import com.example.demo.form.admin.AdminEditForm;
 import com.example.demo.service.admin.AdminService;
 
@@ -33,11 +37,17 @@ public class AdminListController {
 	}
 
 	@GetMapping("/{id}")
-	public String detailAdmin(@PathVariable Long id, Model model) {
+	public String detailAdmin(@PathVariable Long id,
+			@AuthenticationPrincipal UserDetails user,
+			Model model) {
 
 		model.addAttribute("admindetail", adminService.getDetailAdmin(id));
 		model.addAttribute("store", adminService.getStores());
 		model.addAttribute("role", adminService.getRoles());
+
+		AdminEntity loginAdmin = adminService.getAdminByEmail(user.getUsername());
+		model.addAttribute("loginAdmin", loginAdmin);
+
 		return "admin/AdminDetail";
 	}
 
@@ -50,7 +60,7 @@ public class AdminListController {
 		return "admin/AdminEdit";
 	}
 
-	@PostMapping("/update")
+	@PutMapping("/update")
 	public String updateAdmin(@Validated @ModelAttribute("AdminEditForm") AdminEditForm adminEditForm,
 			BindingResult errorResult, Model model) {
 
@@ -62,14 +72,21 @@ public class AdminListController {
 			return "admin/AdminEdit";
 		}
 		adminService.updateAdmin(adminEditForm);
-		return "redirect:/admins/adminList";
+		return "redirect:/admins/list";
 	}
 
-	@GetMapping("/{id}/delete")
+	@DeleteMapping("/{id}/delete")
 	@Transactional
-	public String delete(@PathVariable Long id) {
-		adminService.delete(id);
-		return "redirect:/admins/adminList";
-	}
+	public String delete(@PathVariable Long id,
+			@AuthenticationPrincipal UserDetails user) {
 
+		AdminEntity loginAdmin = adminService.getAdminByEmail(user.getUsername());
+
+		if (loginAdmin.getPermissionId() == null || loginAdmin.getPermissionId() != 1L) {
+			return "redirect:/admins/list";
+		}
+
+		adminService.delete(id);
+		return "redirect:/admins/list";
+	}
 }
